Move existing neural net modules to common one
