+++ tests/nnet/test_sgl_parm.py
@@ -2,9 +2,9 @@ import pytest
 from numpy import array, append, empty, zeros, int64
 from numpy.testing import assert_allclose
 
-from touvlo.nn_clsf import (feed_forward, init_nn_weights,
-                            back_propagation, cost_function,
-                            grad, unravel_params, h)
+from touvlo.nnet.sgl_parm import (feed_forward, init_nn_weights,
+                                  back_propagation, cost_function,
+                                  grad, unravel_params, h)
 
 
 class TestNeuralNetwork:
