+++ touvlo/__init__.py
@@ -2,9 +2,9 @@ import pbr.version
 
 from touvlo import lgx_rg
 from touvlo import lin_rg
-from touvlo import nn_clsf
+from touvlo import nnet
 from touvlo import unsupv
 from touvlo import utils
 
 __version__ = pbr.version.VersionInfo('touvlo').version_string()
-__all__ = ['lgx_rg', 'lin_rg', 'nn_clsf', 'unsupv', 'utils']
+__all__ = ['lgx_rg', 'lin_rg', 'nnet', 'unsupv', 'utils']
