+++ touvlo/nnet/__init__.py
@@ -1,0 +1,4 @@
+from touvlo.nnet import cmpt_grf
+from touvlo.nnet import sgl_parm
+
+__all__ = ['cmpt_grf', 'sgl_parm']
